+++ resources/views/producers/index.blade.php
@@ -13,17 +13,24 @@
 
         <div class="container">
 
-            <a href="{{route('producers.create')}}" class="btn btn-primary a-btn-slide-text">
+            @auth
+                @if(Auth::user()->is_admin)
+                    <a href="{{route('producers.create')}}" class="btn btn-primary a-btn-slide-text">
                     <span class="glyphicon glyphicon-plus" aria-hidden="true"></span>
                     <span><strong>ADD</strong></span>            
-            </a>
+                    </a>
 
-            @if ($message = Session::get('success'))
-                <div class="alert alert-success alert-block">
-                    <button type="button" class="close" data-dismiss="alert">×</button> 
-                    <strong>{{ $message }}</strong>
-                </div>
-            @endif
+                    @if ($message = Session::get('success'))
+                        <div class="alert alert-success alert-block">
+                            <button type="button" class="close" data-dismiss="alert">×</button> 
+                            <strong>{{ $message }}</strong>
+                        </div>
+                    @endif
+
+                @endif
+
+            @endauth
+            
 
             <div class="table-responsive">
                 <table class="table table-striped table-hover">
@@ -31,8 +38,14 @@
                         <tr>    
                         <th>Producers ID</th>
                         <th>Name</th>
-                        <th>Edit</th>
-                        <th>Delete</th>
+
+                        @auth
+                            @if(Auth::user()->is_admin)
+                                <th>Edit</th>
+                                <th>Delete</th>
+                            @endif
+                        @endauth
+                        
                         </tr>
                     </thead>
 
@@ -41,13 +54,18 @@
                             <tr>
                                 <td>{{$producer->producers_id}}</td>
                                 <td><a href="{{route('producers.show',$producer->producers_id)}}">{{$producer->name}}</a></td>
-                                <td align="left"><a href="{{ route('producers.edit',$producer->producers_id) }}"> <i class="fa fa-pencil-square-o" aria-hidden="true" style="font-size:24px" ></a></i></td>
-                                <td align="left">
-                                    {!! Form::open(array('route' => array('producers.destroy', $producer->producers_id),'method'=>'DELETE')) !!}
-                                        <button><i class="fa fa-trash-o" style="font-size:24px; color:red" ></i></button>
-                                    {!! Form::close() !!}
-                                </td>
-                                    
+                                
+                                @auth
+                                    @if(Auth::user()->is_admin)
+                                        <td align="left"><a href="{{ route('producers.edit',$producer->producers_id) }}"> <i class="fa fa-pencil-square-o" aria-hidden="true" style="font-size:24px" ></a></i></td>
+                                        <td align="left">
+                                            {!! Form::open(array('route' => array('producers.destroy', $producer->producers_id),'method'=>'DELETE')) !!}
+                                                <button><i class="fa fa-trash-o" style="font-size:24px; color:red" ></i></button>
+                                            {!! Form::close() !!}
+                                        </td>
+                                    @endif
+                                @endauth
+
                             </tr>
 
                         @endforeach
